+++ src/theme/progress-bar.js
@@ -2,6 +2,6 @@ const { lightness } = require('../base/util');
 
 exports.progressBar = function(color) {
     return {
-        'progressBar.background': color.blue,
+        'progressBar.background': color.cyan,
     }
 }