+++ src/theme/quick-picker.js
@@ -2,13 +2,13 @@ const { lightness } = require('../base/util');
 
 exports.quickPicker = function(color) {
     return {
-        // 'pickerGroup.border': '',
+        'pickerGroup.border': lightness(color.bg, -2),
         'pickerGroup.foreground': color.blue,
-        // 'quickInput.background': '',
-        // 'quickInput.foreground': '',
-        // 'quickInputList.focusBackground': '',
-        // 'quickInputList.focusForeground': '',
-        // 'quickInputList.focusIconForeground': '',
-        // 'quickInputTitle.background': '',
+        'quickInput.background': lightness(color.bg, -2),
+        'quickInput.foreground': color.fg,
+        'quickInputList.focusBackground': lightness(color.bg, 2),
+        'quickInputList.focusForeground': color.fgBright,
+        'quickInputList.focusIconForeground': color.cyan,
+        'quickInputTitle.background': lightness(color.bg, 2),
     }
 }